sub_service/auth_service/requirements.txt
Flask==1.1.2
Flask-Script==2.0.6
Flask-Cors==3.0.9
requests==2.25.0
gunicorn==20.0.4
PyJWT==1.7.1